refactor: clean things up
