+++ server/Server.API/Data/DbOptions.cs
@@ -1,0 +1,7 @@
+namespace Server.API.Data;
+
+record DbOptions
+{
+  public string ConnectionString { get; init; } = string.Empty;
+  public string DatabaseName { get; init; } = string.Empty;
+}
