+++ server/Server.API/Extensions/HttpContextExtensions.cs
@@ -1,8 +1,6 @@
-using System.Security.Claims;
-
 namespace Server.API.Extensions;
 
-public static class HttpContextExtensions
+static class HttpContextExtensions
 {
   public static string? GetUserId(this HttpContext context)
   {
