--- server/Server.API/Models/DbOptions.cs
@@ -1,7 +1,0 @@
-namespace Server.API.Models;
-
-class DbOptions
-{
-  public string ConnectionString { get; set; } = string.Empty;
-  public string DatabaseName { get; set; } = string.Empty;
-}
