+++ server/Server.API/Models/JwtOptions.cs
@@ -1,8 +1,8 @@
 namespace Server.API.Models;
 
-class JwtOptions
+record JwtOptions
 {
-  public string Secret { get; set; } = string.Empty;
-  public string Issuer { get; set; } = string.Empty;
-  public string Audience { get; set; } = string.Empty;
+  public string Secret { get; init; } = string.Empty;
+  public string Issuer { get; init; } = string.Empty;
+  public string Audience { get; init; } = string.Empty;
 }
