+++ server/Server.API/Usings.cs
@@ -1,5 +1,6 @@
 global using System.Text;
 global using System.Reflection;
+global using System.Security.Claims;
 
 global using FluentResults;
 
@@ -10,10 +11,12 @@ global using Microsoft.Extensions.Options;
 global using Microsoft.IdentityModel.Tokens;
 global using Microsoft.AspNetCore.Mvc;
 
+
 global using MongoDB.Bson;
 global using MongoDB.Bson.Serialization.Attributes;
 global using MongoDB.Driver;
 
-global using Server.API.Dtos;
-global using Server.API.Interfaces;
+global using Server.API.Controllers;
 global using Server.API.Models;
+global using Server.API.Data;
+global using Server.API.Extensions;
